Sidebar navigation added to dashboard
